feat(): Improve mikro orm serializer performances
